+++ pkg/dbt_cloud/athena_credential.go
@@ -27,7 +27,7 @@ type AthenaUnencryptedCredentialDetails struct {
 // AthenaCredentialData represents the data returned by the API for an Athena credential
 type AthenaCredentialData struct {
 	ID                           *int                               `json:"id"`
-	AccountID                    int                                `json:"account_id"`
+	AccountID                    int64                               `json:"account_id"`
 	ProjectID                    int                                `json:"project_id"`
 	Type                         string                             `json:"type"`
 	State                        int                                `json:"state"`
